\name{lab.util-package}
\alias{lab.util-package}
\alias{lab.util}
\docType{package}
\title{
Utility fucntions for the Psychoacoustics Lab.
}
\description{
This provides a growing library of useful functions for doing data
anlaysis in our lab using R.
}
\details{
\tabular{ll}{
Package: \tab lab.util\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-25\cr
License: \tab What license is it under?\cr
}
This is a hodgepodge of functions which are not yet well documented.
}
\author{
David Little
}
\keyword{ package }
\seealso{
## ~~ Optional links to other man pages, e.g. ~~
## ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
## ~~ simple examples of the most important functions ~~
}
